% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instaquote.R
\name{insta_text}
\alias{insta_text}
\title{Generate HTML for quotes}
\usage{
insta_text(title, subtitle = NULL, title_size = 30, subtitle_size = 12)
}
\arguments{
\item{title}{Title}

\item{subtitle}{Subtitle (optional)}

\item{title_size}{Font size of title. Default:30}

\item{subtitle_size}{}
}
\value{

}
\description{
Generate HTML for quotes
}
\examples{
insta_text("hi there!")
}
