% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instaquote.R
\name{instaquote}
\alias{instaquote}
\title{Generate a ggplot of a quote}
\usage{
instaquote(
  text,
  bg_color = "#56B4E9",
  text_color = "white",
  caption = "gospelanalysis.com"
)
}
\arguments{
\item{text}{Text you want to pass in. use generate_text().}

\item{bg_color}{Plot background color, default: ggthemes::palette_pander(n = 1)}

\item{text_color}{Color of text. default: white.}
}
\value{
ggplot2 object
}
\description{
Uses `ggtext::geom_textbox()` and `ggplot2::theme_void()` to generate
a simple quote. Quotes can be posted to social media or within rmarkdown.
}
\examples{
txt <- insta_text("hello world")
instaquote(txt)
}
